+++ temp/others/rollinghash.cpp
@@ -91,7 +91,7 @@ template<typename T> vl get_hash(vector<T>& s) { vl h(s.size() + 1, 0); rep(i, s
 vl get_hash(string& s) { vl h(s.size() + 1, 0); rep(i, s.size()) { h[i + 1] = (P_B * h[i] + s[i]) % P_M; } return h; }
 
 ll sub_hash(vl& h, ll l, ll r) {
-  assert(0 <= l && r <= h.size());
+  assert(0 <= l && r <= (ll)h.size());
   ll ret = h[r] - (b_fact(r - l) * h[l] % P_M); if (ret < 0) ret += P_M;
   return ret;
 }
